% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_missing.R
\name{apply_missing}
\alias{apply_missing}
\title{Transforms strings to NA}
\usage{
apply_missing(df, missing_strings = c("", "NA", "NULL"))
}
\arguments{
\item{df}{data.frame with values that should be NA}

\item{missing_strings}{strings that should be replaced by NA}
}
\value{
data.frame with strings replaced by NA
}
\description{
Transforms strings to NA
}
