% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_missing.R
\name{find_missing}
\alias{find_missing}
\alias{remove_missing}
\title{Find missing cells}
\usage{
find_missing(data, tolerance = 0.9)

remove_missing(data, tolerance = 0.9)
}
\arguments{
\item{data}{data.frame to be checked}

\item{tolerance}{maximum fraction of missing
values allowed in a column.}

\item{data}{data.frame to be checked}

\item{tolerance}{maximum fraction of missing
values allowed in a column.}
}
\value{
summary of rows and columns that
could be removed

summary of rows and columns that

summary of rows and columns that
could be removed
}
\description{
Identifies how many rows would be
removed if columns with number of missing
above tolerance are removed.

Removes rows with missing values after
removing columns with fraction of missing
values above tolerance.
}
