% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{write_compressed}
\alias{write_compressed}
\title{Saves DataFrame to compressed file}
\usage{
write_compressed(df, path, compress = "xz")
}
\arguments{
\item{df}{DataFrame to be saved}

\item{path}{String. Path to file to be saved without extension}

\item{compress}{String. Type of compression. Only using .xz or .gz. Default is .xz}
}
\description{
Saves a DataFrame to compressed .xz or .gz file. Default is .xz
}
